rustc_target: Fix RISC-V "V" dependency
In general, all vector extensions depend on corresponding scalar arithmetic
extensions (e.g. Zve32f depends on "F" because it supports single-precision
FP arithmetic while does not depend on "D" because it is not obligated
to support double-precision FP arithmetic).

Because the "V" extension supports double-precision FP arithmetic, it must
depend on the "D" extension.