rustdoc: Allow multiple references to a single footnote
Multiple references to a single footnote is not prohibited by rustdoc but
causes multiple "sup" elements with the same "id" attribute, which is
invalid per the HTML specification.

However, such references are helpful on certain cases and actually tested
in tests/rustdoc/footnote-reference-in-footnote-def.rs.

This commit keeps track of the number of references per footnote and gives
unique ID per footnote reference and emits back links to *all* footnote
references as "↩" (return symbol) plus a numeric list in superscript,
assuming that all references to a footnote are rendered.

Also note that, this commit is constructed so that there's no actual HTML
changes unless there are multiple references to a single footnote.